Update for the Android version which presents a popup to the user to ask for permissions.
This is the new security model in Android API level 23, to which I bumped the app without much thought. Unfortunately, the solution needs some Java code as a workaround, as Qt is still working on a nice way to do this. This should arrive in Qt 5.10 later this year, at which time it should probably adopted as the proper solution.